% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_evaluation.R
\name{acc_intercept}
\alias{acc_intercept}
\title{accuracy intercept model}
\usage{
acc_intercept(elnet_result)
}
\arguments{
\item{elnet_result}{result of el_net function}
}
\description{
calculates the accuracy of an intercept model
}
