% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{calc_acc_elnet}
\alias{calc_acc_elnet}
\title{calculate accuracy for one elastic net model}
\usage{
calc_acc_elnet(pred, y)
}
\arguments{
\item{pred}{matrix with predictions}

\item{y}{real y of predicitons}
}
\description{
calculates accuracy for one elastic net model for several lambda values, returns accuracy for every lambda
}
