% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{calc_el_net}
\alias{calc_el_net}
\title{elastic net model for one alpha}
\usage{
calc_el_net(data_train, data_test, vars, alpha = 0, metric = "accuracy", ...)
}
\arguments{
\item{data_train}{training data}

\item{data_test}{test data}

\item{vars}{which variables should be used? If null, columns containing connectivity data are automatically extracted}

\item{alpha}{alpha value for glmnet. 0 = Lasso, 1 = Ridge}

\item{metric}{which metric is used to find the best lambda? !!! ToDo: add more metrics, think about advantages/disadvantages}

\item{...}{Additional parameters for glmnet function, e.g. nlambda}
}
\description{
calculates elastic net model for one alpha value and chooses best model according to some metric (returns list with information)
}
