% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{el_net}
\alias{el_net}
\title{elastic net model}
\usage{
el_net(data, vars = NULL, test_IDs, alpha, y_0, y_1, metric = "accuracy", ...)
}
\arguments{
\item{data}{dataset}

\item{vars}{which variables should be used? If null, columns containing connectivity data are automatically extracted}

\item{test_IDs}{IDs for test data}

\item{alpha}{alpha value for glmnet. 0 = Lasso, 1 = Ridge}

\item{y_0}{which levels of data$prmdiag use as 0}

\item{y_1}{which levels of data$prmdiag use as 1}

\item{metric}{which metric is used to find the best lambda? !!! ToDo: add more metrics, think about advantages/disadvantages}

\item{...}{Additional parameters for glmnet function, e.g. nlambda}

\item{data_source}{which dataset is it? (in case there have to be made adjustments)}
}
\description{
elastic net model
}
