% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{evaluation_elnet}
\alias{evaluation_elnet}
\title{evaluation for one elastic net model}
\usage{
evaluation_elnet(pred, y)
}
\arguments{
\item{pred}{matrix with predictions}

\item{y}{real y of predicitons}
}
\description{
calculates accuracy, auc for one elastic net model for several lambda values
}
