% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_evaluation.R
\name{get_auc_elnet}
\alias{get_auc_elnet}
\title{get auc for elastic net model}
\usage{
get_auc_elnet(elnet_result, ind_alpha, ind_lambda, test = TRUE)
}
\arguments{
\item{elnet_result}{result of el_net function}

\item{ind_alpha}{index of alpha that should be used}

\item{ind_lambda}{index of lambda value that should be used}

\item{test}{logical: use test or train data}
}
\description{
get auc for elastic net model
}
