% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{names_conn}
\alias{names_conn}
\title{extract names of connectivity variables}
\usage{
names_conn(colnames)
}
\arguments{
\item{colnames}{colnames of dataset}
}
\description{
extract names of connectivity variables
}
