% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_preprocessing.R
\name{prep_y}
\alias{prep_y}
\title{prepare outcome variable}
\usage{
prep_y(data, y_0, y_1)
}
\arguments{
\item{data}{dataset}

\item{y_0}{which levels of data$prmdiag use as 0}

\item{y_1}{which levels of data$prmdiag use as 1}
}
\description{
returns dataset with column y added
}
