% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{result_table_elnet}
\alias{result_table_elnet}
\title{get result table based on metric for el_net result}
\usage{
result_table_elnet(elnet_result)
}
\arguments{
\item{elnet_result}{result of el_net function}
}
\description{
returns best values (mretric, lambda) for each alpha
}
