% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_preprocessing.R
\name{train_test_data}
\alias{train_test_data}
\title{prepare test and training data}
\usage{
train_test_data(data, train_IDs, test_IDs = NULL)
}
\arguments{
\item{data}{dataset}

\item{train_IDs}{IDs for training data}

\item{test_IDs}{IDs for test data, if null, use all data that is not in training data}
}
\description{
returns list with train and test data
}
