% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{calc_el_net}
\alias{calc_el_net}
\title{elastic net model for one alpha}
\usage{
calc_el_net(data_train, data_test, alpha = 0, target_diag, ...)
}
\arguments{
\item{data_train}{training data}

\item{data_test}{test data}

\item{alpha}{alpha value for glmnet. 0 = Lasso, 1 = Ridge}

\item{target_diag}{diagnosis as target variable (logistic regression) or linear model for MEM-score}

\item{...}{Additional parameters for glmnet function, e.g. nlambda}
}
\description{
calculates elastic net model for one alpha value (returns list with information and evaluation)
}
