% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{el_net}
\alias{el_net}
\title{elastic net model}
\usage{
el_net(
  test,
  train,
  vars = NULL,
  alpha = seq(0, 1, by = 0.1),
  y_0 = NULL,
  y_1 = NULL,
  data_source = "DELCODE",
  target_diag = TRUE,
  option = "standard",
  ...
)
}
\arguments{
\item{test}{test data}

\item{train}{training data}

\item{vars}{which variables should be used? If null, columns containing connectivity data are automatically extracted}

\item{alpha}{alpha values for glmnet. 0 = Lasso, 1 = Ridge}

\item{y_0}{which levels of data$prmdiag use as 0}

\item{y_1}{which levels of data$prmdiag use as 1}

\item{data_source}{which dataset is it? (in case there have to be made adjustments)}

\item{target_diag}{diagnosis as target variable (logistic regression) or linear model for MEM-score}

\item{option}{standard: standard regression model, interactions: all two way interactions, quadratic: all quadratic terms (x1^2, x2^2, but no interactions), abs: absolute value of variables, squ: squared value of variables}

\item{...}{Additional parameters for glmnet function, e.g. nlambda}
}
\description{
calculates elastic net model for several alpha values
}
