% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_confMatrix_elnet}
\alias{get_confMatrix_elnet}
\title{get confusion matrix for best elastic net model}
\usage{
get_confMatrix_elnet(elnet_result)
}
\arguments{
\item{elnet_result}{result of el_net function}
}
\description{
get confusion matrix for best elastic net model
}
