% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_preprocessing.R
\name{names_conn}
\alias{names_conn}
\title{returns names of connectivity variables}
\usage{
names_conn(colnames)
}
\arguments{
\item{colnames}{colnames of dataset}
}
\description{
returns names of connectivity variables
}
