% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_evaluation.R
\name{plot_matrix_coeffs}
\alias{plot_matrix_coeffs}
\title{plot matrix of coefficients}
\usage{
plot_matrix_coeffs(beta)
}
\arguments{
\item{beta}{coefficients of one model}
}
\description{
plot matrix of coefficients
}
