% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_evaluation.R
\name{plot_matrix_coeffs}
\alias{plot_matrix_coeffs}
\title{plot matrix of coefficients, probably doesn't work for when beta contains interaction terms!}
\usage{
plot_matrix_coeffs(beta, regions_dat = NULL)
}
\arguments{
\item{beta}{coefficients of one model}

\item{regions_dat}{data on regions of labels. First column contains labels, second column contains regions}
}
\description{
plot matrix of coefficients, probably doesn't work for when beta contains interaction terms!
}
